-View For Seats Done -Dynamicly show seats according the number of rows and columns in flight chosen

+++ Domain/Models/Ticket.cs
@@ -20,7 +20,7 @@ namespace Domain.Models
         [ForeignKey("Flight")]
         public Guid FlightIdFK { get; set; }
         public string Passport { get; set; }
-        public int PricePaid { get; set; }
+        public double PricePaid { get; set; }
         public bool Cancelled { get; set; }
     }
 }
